Routes, Protections, Login and Doc are finalized.

+++ src/modules/users/users.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repositories/user.repository';
@@ -9,6 +9,10 @@ export class UsersService {
   constructor(private usersRepository: UsersRepository){}
   
   async create(createUserDto: CreateUserDto) {
+    const findUser = await this.usersRepository.findByEmail(createUserDto.email);
+    if (findUser) {
+      throw new ConflictException('User already exists');
+    } 
     const user = await this.usersRepository.create(createUserDto);
     return user;
   }
@@ -20,15 +24,31 @@ export class UsersService {
 
   async findOne(id: string) {
     const user = await this.usersRepository.findOne(id);
+    if (!user){
+      throw new NotFoundException('User not found');
+    } 
+    return user
+  }
+
+  async findByEmail(email: string) {
+    const user = await this.usersRepository.findByEmail(email);
     return user
   }
 
   async update(id: string, updateUserDto: UpdateUserDto) {
+    const findUser = await this.usersRepository.findOne(id);
+    if (!findUser){
+      throw new NotFoundException('User not found');
+    } 
     const user = await this.usersRepository.update(id, updateUserDto);
     return user
   }
 
   async remove(id: string) {
+        const findUser = await this.usersRepository.findOne(id);
+    if (!findUser){
+      throw new NotFoundException('User not found');
+    } 
     await this.usersRepository.delete(id);
     return;
   }
